update project with hex arch
