+++ gerenciadorLivros/pom.xml
@@ -85,6 +85,11 @@
       		<artifactId>springdoc-openapi-ui</artifactId>
 			<version>1.6.6</version>
    		</dependency>
+		<dependency>
+			<groupId>co.elastic.clients</groupId>
+			<artifactId>elasticsearch-java</artifactId>
+			<version>8.1.3</version>
+		</dependency>
 
 	</dependencies>
