+++ gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/core/adapters/catalog/elasticSearch/RestService.java
@@ -1,0 +1,27 @@
+package com.deborateste.gerenciadorLivros.core.adapters.catalog.elasticSearch;
+
+import com.deborateste.gerenciadorLivros.core.domain.model.Livro;
+import java.io.IOException;
+import java.util.List;
+
+public class RestService {
+
+    public static void post(Livro livro) throws IOException {
+        // TO DO
+    }
+
+    public static void deleteAll() throws IOException {
+        // TO DO
+    }
+
+    public static List<Livro> search(String searchString) throws IOException {
+        // TO DO
+        return null;
+    }
+
+    public static List<Livro> getList() throws IOException {
+        // TO DO
+        return null;
+    }
+
+}
