+++ gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/core/adapters/repository/GenreDAO.java
@@ -1,0 +1,9 @@
+package com.deborateste.gerenciadorLivros.core.adapters.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.deborateste.gerenciadorLivros.core.domain.model.Genero;
+
+public interface GenreDAO extends CrudRepository<Genero, Integer>{
+
+}
