--- gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/core/domain/Genero.java
@@ -1,51 +1,0 @@
-package com.deborateste.gerenciadorLivros.model;
-
-import java.util.List;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@Entity
-@Table(name="generos")
-public class Genero {
-	
-	@Id
-	@Column(name="id", nullable=false)
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer idGenero;
-	
-	@Column(name="nome", nullable=false, length=100)
-	private String nome;
-	
-	@OneToMany(mappedBy="genero")
-	@JsonIgnoreProperties("genero")
-	private List<Livro> listaLivros;
-	
-	
-	public Integer getIdGenero() {
-		return idGenero;
-	}
-	public void setIdGenero(Integer idGenero) {
-		this.idGenero = idGenero;
-	}
-	public String getNome() {
-		return nome;
-	}
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-	public List<Livro> getListaLivros() {
-		return listaLivros;
-	}
-	public void setListaLivros(List<Livro> listaLivros) {
-		this.listaLivros = listaLivros;
-	}
-	
-}
