--- gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/core/domain/Livro.java
@@ -1,71 +1,0 @@
-package com.deborateste.gerenciadorLivros.model;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@Entity
-@Table(name="livros")
-public class Livro {
-	
-	@Id
-	@Column(name="id", nullable=false)
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer idLivro;
-	
-	@Column(name="nome", nullable=false)
-	private String titulo;
-	
-	@Column(name="autor", nullable=false)
-	private String autor;
-	
-	@Column(name="num_paginas")
-	private Integer numPaginas;
-	
-	@ManyToOne
-	@JoinColumn(name="id_genero")
-	@JsonIgnoreProperties("listaLivros")
-	private Genero genero;
-	
-	
-	public Integer getIdLivro() {
-		return idLivro;
-	}
-	public void setIdLivro(Integer idLivro) {
-		this.idLivro = idLivro;
-	}
-	public String getTitulo() {
-		return titulo;
-	}
-	public void setTitulo(String titulo) {
-		this.titulo = titulo;
-	}
-	public String getAutor() {
-		return autor;
-	}
-	public void setAutor(String autor) {
-		this.autor = autor;
-	}
-	public Integer getNumPaginas() {
-		return numPaginas;
-	}
-	public void setNumPaginas(Integer numPaginas) {
-		this.numPaginas = numPaginas;
-	}
-	public Genero getGenero() {
-		return genero;
-	}
-	public void setGenero(Genero genero) {
-		this.genero = genero;
-	}
-	
-	
-
-}
