+++ gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/core/domain/model/Usuario.java
@@ -1,0 +1,70 @@
+package com.deborateste.gerenciadorLivros.core.domain.model;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@Table(name="usuarios")
+public class Usuario {
+	@Id
+	@Column(name="id", nullable=false)
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	private Integer id;
+	
+	@Column(name="nome", nullable=false)
+	private String nome;
+	
+	@Column(name="email", nullable=false, unique=true)
+	private String email;
+	
+	@Column(name="username", nullable=false, unique=true)
+	private String username;
+	
+	@Column(name="senha", nullable=false)
+	private String senha;
+
+	public Integer getId() {
+		return id;
+	}
+
+	public void setId(Integer id) {
+		this.id = id;
+	}
+
+	public String getUsername() {
+		return username;
+	}
+
+	public void setUsername(String username) {
+		this.username = username;
+	}
+
+	public String getSenha() {
+		return senha;
+	}
+
+	public void setSenha(String senha) {
+		this.senha = senha;
+	}
+
+	public String getNome() {
+		return nome;
+	}
+
+	public void setNome(String nome) {
+		this.nome = nome;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+	
+}
