+++ gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/core/domain/model/dto/UserLoginDTO.java
@@ -1,0 +1,20 @@
+package com.deborateste.gerenciadorLivros.core.domain.model.dto;
+
+public class UserLoginDTO {
+	private String username;
+	private String senha;
+	public String getUsername() {
+		return username;
+	}
+	public void setUsername(String username) {
+		this.username = username;
+	}
+	public String getSenha() {
+		return senha;
+	}
+	public void setSenha(String senha) {
+		this.senha = senha;
+	}
+	
+
+}
