+++ gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/core/domain/ports/IBookCatalog.java
@@ -1,0 +1,10 @@
+package com.deborateste.gerenciadorLivros.core.domain.ports;
+
+import com.deborateste.gerenciadorLivros.core.domain.model.Livro;
+import java.util.List;
+
+public interface IBookCatalog {
+    List<Livro> searchBook(String search);
+    Livro addOne(Livro book);
+    void deleteAll();
+}
