--- gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/features/book/repository/LivroDAO.java
@@ -1,9 +1,0 @@
-package com.deborateste.gerenciadorLivros.dao;
-
-//import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-
-import com.deborateste.gerenciadorLivros.model.Livro;
-
-public interface LivroDAO extends CrudRepository<Livro, Integer>{
-}
