--- gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/features/gender/repository/GeneroDAO.java
@@ -1,9 +1,0 @@
-package com.deborateste.gerenciadorLivros.dao;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.deborateste.gerenciadorLivros.model.Genero;
-
-public interface GeneroDAO extends CrudRepository<Genero, Integer>{
-
-}
