--- gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/features/gender/service/IGeneroService.java
@@ -1,9 +1,0 @@
-package com.deborateste.gerenciadorLivros.service;
-
-import java.util.List;
-
-import com.deborateste.gerenciadorLivros.model.Genero;
-
-public interface IGeneroService {
-	public List<Genero> recuperarTodos();
-}
