--- gerenciadorLivros/src/main/java/com/deborateste/gerenciadorLivros/features/user/repository/UserDAO.java
@@ -1,10 +1,0 @@
-package com.deborateste.gerenciadorLivros.dao;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.deborateste.gerenciadorLivros.model.Usuario;
-
-public interface UserDAO extends CrudRepository<Usuario, Integer>{
-	public Usuario findByUsername(String username);
-	public Usuario findByEmail(String email);
-}
